refactor(auth): some renaming
