+++ src/components/layouts/auth/AuthLayout.styles.ts
